parte3/requirements.txt
# Dependências para Parte 3 - Banco de Dados
mysql-connector-python>=8.0.33
